Fixing Small File upload using Chunks
An issue arises on small file uploads where the chunk has been read, and the buffer has been set... and then we finishUpload which causes the file to be duplicated in content.

This is due to trying to re-use a single finishUpload call at the end. A quick fix is to reset the buffer on the small upload first.

We should revisit this code in v5.